refactor: FixedNumberGenerator 패키지 변경
