+++ src/test/java/racingCar/domains/CarTest.java
@@ -3,6 +3,7 @@ package racingCar.domains;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingCar.utils.FixedNumberGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
